src/tests/requirements.txt
pytest==5.2.2
python-telegram-bot==11.1.0
SQLAlchemy==1.3.10